Removes test_fb_misc.py. Minor adjustments (formatting).
