+++ test/test_dialect.py
@@ -23,8 +23,33 @@ from sqlalchemy.testing.assertions import AssertsExecutionResults
 from sqlalchemy.testing.assertions import eq_
 
 
+class ConnectionTest(fixtures.TablesTest):
+    def test_is_disconnect(self):
+        try:
+            with testing.db.begin() as first_conn:
+                con1_id = first_conn.exec_driver_sql(
+                    "SELECT CURRENT_CONNECTION FROM rdb$database"
+                ).scalar()
+
+                with testing.db.begin() as second_conn:
+                    # Kills first_conn
+                    second_conn.exec_driver_sql(
+                        "DELETE FROM mon$attachments WHERE mon$attachment_id = ?",
+                        (con1_id,),
+                    )
+
+                # Attemps to read from first_conn
+                first_conn.exec_driver_sql(
+                    "SELECT CURRENT_CONNECTION FROM rdb$database"
+                )
+
+                assert False
+        except Exception as err:
+            eq_(testing.db.dialect.is_disconnect(err.orig, None, None), True)
+
+
 #
-# Tests from postgresql/test_compiler.py
+# Tests from postgresql/test_dialect.py
 #
