+++ test/test_query.py
@@ -15,6 +15,7 @@ from sqlalchemy import select
 from sqlalchemy import Sequence
 from sqlalchemy import String
 from sqlalchemy import Table
+from sqlalchemy import testing
 from sqlalchemy import text
 from sqlalchemy import Time
 from sqlalchemy.testing import assert_raises
@@ -27,6 +28,45 @@ from sqlalchemy.testing import requires
 from sqlalchemy.testing import skip_if
 
 from sqlalchemy_firebird.types import _FBInterval
+
+
+class QueryTest(fixtures.TestBase):
+    @testing.provide_metadata
+    def test_strlen(self, connection):
+        metadata = self.metadata
+
+        t = Table(
+            "t1",
+            metadata,
+            Column("id", Integer, Sequence("t1idseq"), primary_key=True),
+            Column("name", String(10)),
+        )
+        metadata.create_all(testing.db)
+        connection.execute(t.insert().values(dict(name="dante")))
+        connection.execute(t.insert().values(dict(name="alighieri")))
+        eq_(
+            connection.execute(
+                select(func.count(t.c.id)).where(func.length(t.c.name) == 5)
+            ).scalar(),
+            1,
+        )
+
+    def test_percents_in_text(self, connection):
+        for expr, result in (
+            (text("select '%' from rdb$database"), "%"),
+            (text("select '%%' from rdb$database"), "%%"),
+            (text("select '%%%' from rdb$database"), "%%%"),
+            (
+                text("select 'hello % world' from rdb$database"),
+                "hello % world",
+            ),
+        ):
+            eq_(connection.scalar(expr), result)
+
+
+#
+# Tests from postgresql/test_query.py
+#
 
 
 class FunctionTypingTest(fixtures.TestBase, AssertsExecutionResults):
